modified data members in Account and Transfer

+++ 09-Capstone/student/java/tenmo-server/src/main/java/com/techelevator/tenmo/model/Account.java
@@ -4,8 +4,8 @@ import java.math.BigDecimal;
 
 public class Account {
 
-	private long accountId;
-	private long userId;
+	private int accountId;
+	private int userId;
 	private BigDecimal balance;
 	
 	
@@ -13,7 +13,7 @@ public class Account {
 		
 	}
 	
-	public Account(long accountId, long userId, BigDecimal balance) {
+	public Account(int accountId, int userId, BigDecimal balance) {
 		this.accountId = accountId;
 		this.userId = userId;
 		this.balance = balance;
@@ -23,7 +23,7 @@ public class Account {
 		return accountId;
 	}
 
-	public void setAccountId(long accountId) {
+	public void setAccountId(int accountId) {
 		this.accountId = accountId;
 	}
 
@@ -31,7 +31,7 @@ public class Account {
 		return userId;
 	}
 
-	public void setUserId(long userId) {
+	public void setUserId(int userId) {
 		this.userId = userId;
 	}
